News feed test modifications according to PR comments
